// api/seller/client/models/GetAllTemplateResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type GetAllTemplateResponse = {
  variantOption: Array<{
    total: number;
    name: string;
  }>;
  file: string;
  isActive: boolean;
  minSellingPrice: number;
  description: string | null;
  podTechniqueId: number;
  podCategoryId: number;
  podProductTypeId: number;
  name: string;
  id: number;
};